Refactored search results, added models
