+++ common/djangoapps/search/sorting.py
@@ -8,10 +8,12 @@ def sort(data_list, sorting):
 
 
 def relevance_sort(data_list):
-    sorting = lambda entry: entry[4]
-    return sorted(data_list, key=sorting)
+    sorting = lambda entry: entry.data.get("score", 0)
+    ascending = sorted(data_list, key=sorting)
+    ascending.reverse()  # Now descending
+    return ascending
 
 
 def alphabetical_sort(data_list):
-    sorting = lambda entry: " ".join(entry[0].split()[2:]).lower()
+    sorting = lambda entry: " ".join(entry.data.get("display_name", "")).lower()
     return sorted(data_list, key=sorting)
